Testing2

+++ app/controllers/admincontrols.php
@@ -10,17 +10,18 @@ use function PHPSTORM_META\type;
                 $this->redirect("CommonControls/loadLoginView");
             }
 
+            $customer = new Customer();
             $systemuser = new Systemuser();
-            $producorder = new ProductOrder();
-            $productorderline = new ProductOrderLine();
+            $outlet = new Outlet();
             $productitem = new ProductItem();
 
-            $data = $systemuser->where("UserName", $_SESSION["USER"]->UserName);
-            $producorderdata = $producorder->findall();
-            $productorderlinedata = $productorderline->findall();
-            $productitemdata = $productitem->findall();
+            $customercount = $customer->getCustomerCount();
+            $systemusercount = $systemuser->getSystemuserCount();
+            $outletcount = $outlet->getOutletCount();
+            $productitemcount = $productitem->countProductItemsGroupByCategory();
 
-            echo $this->view("admin/admindash",[ "data" => $data, "producorderdata" => $producorderdata, "productorderlinedata" => $productorderlinedata, "productitemdata" => $productitemdata]);
+            
+            echo $this->view("admin/admindash", ["customercount" => $customercount, "systemusercount" => $systemusercount, "outletcount" => $outletcount, "productitemcount" => $productitemcount]);
         }
 
         //Item functions
@@ -301,7 +302,8 @@ use function PHPSTORM_META\type;
             $arr["Address"] = $_POST["Address"];
             $arr["Role"] = $_POST["Role"];
             $arr["UserName"] = $_POST["UserName"];
-            $arr["Password"] = $_POST["Password1"];
+            $password = $_POST["Password1"];
+            $arr["Password"] = password_hash($password, PASSWORD_DEFAULT);
         
             // Validate NIC
             $patternNIC = '/^\d{9}[vV]$|^\d{12}$/';
@@ -317,10 +319,8 @@ use function PHPSTORM_META\type;
                 return;
             }
         
-            // Validate Email
             $patternEmail = '/^\w+@[a-zA-Z_]+?\.[a-zA-Z]{2,3}$/';
             if (!preg_match($patternEmail, $arr["Email"])) {
-            // Check if email already in use
             $existingUser = $systemuser->where("Email", $arr["Email"]);
             if ($existingUser) {
                 echo $this->view("admin/AddUserView", ["error" => "Email already in use"]);
@@ -330,10 +330,8 @@ use function PHPSTORM_META\type;
                 return;
             }
         
-            // Validate Contact Number
             $patternContactNo = '/^\d{10}$/';
             if (!preg_match($patternContactNo, $arr["contactNo"])) {
-            // Check if contact number already in use
             $existingUser = $systemuser->where("contactNo", $arr["contactNo"]);
             if ($existingUser) {
                 echo $this->view("admin/AddUserView", ["error" => "Contact number already in use"]);
@@ -343,15 +341,20 @@ use function PHPSTORM_META\type;
                 echo $this->view("admin/AddUserView", ["error" => "Invalid contact number"]);
                 return;
             }
+
+            $existingUserName = $systemuser->where("UserName", $arr["UserName"]);
+            if ($existingUserName) {
+                echo $this->view("admin/AddUserView", ["error" => "Username already in use"]);
+                return;
+            }
         
-            // Validate Role
             $validRoles = ["admin", "billingclerk", "outletmanager", "productionmanager", "receptionist", "storemanager"];
             if (!in_array($arr["Role"], $validRoles)) {
                 echo $this->view("admin/AddUserView", ["error" => "Invalid Role"]);
                 return;
             }
         
-            // Generate EmployeeNo based on Role
+            
             switch ($arr["Role"]) {
                 case "admin":
                     $C = "AD";
@@ -372,7 +375,7 @@ use function PHPSTORM_META\type;
                     $C = "SM";
                     break;
                 default:
-                    echo $this->view("admin/AddUserView", ["error" => "Invalid Role"]);
+                    echo $this->view("admin/AddUser", ["error" => "Invalid Role"]);
                     return;
             }
         
@@ -385,8 +388,10 @@ use function PHPSTORM_META\type;
             $arr["EmployeeNo"] = $C.$max_user_id;
         
             $systemuser->insert($arr);
-        
+            $mail = new Mail();
+            $mail->sendMail($arr["Email"], "User Account Created", "Your account has been created you should change your password on 1st login.<br> username :  <b>".$arr["UserName"]."</b> <br> password : <b>".$password." </b>.");
             $this->redirect(BASE_URL."AdminControls/loadUsersView");
+        
         }
         
         function loadUsersView(){
@@ -415,7 +420,8 @@ use function PHPSTORM_META\type;
             }
 
             $systemuser = new Systemuser();
-            $systemuser->delete($id,"UserID");
+            $systemuser->update($id, "UserID", ["ActiveState" => "Inactive"]);
+            
             $this->redirect(BASE_URL."AdminControls/loadUsersView");
         }
 
@@ -587,6 +593,17 @@ use function PHPSTORM_META\type;
             }
             echo $this->view("admin/systemusers", [ "users" => $users]);
         }
+
+        public function ResetPassword($username){
+            $generatedpassword = uniqid();
+            $passwordhashed = password_hash($generatedpassword, PASSWORD_DEFAULT);
+            $systemuser = new Systemuser();
+            $founduser = $systemuser->where("UserName", $username);
+            $systemuser->update($username, "UserName", ["Password" => $passwordhashed, "ActiveState" => "FirstLogin"]);
+            $mail = new Mail();
+            $mail->sendMail($founduser[0]->Email, "Password Reset", "Your password has been reset. Your new password is : ".$generatedpassword);
+            $this->redirect(BASE_URL."AdminControls/loadUsersView");
+        }
         
         //view add user page
         function AddUser(){
@@ -610,17 +627,17 @@ use function PHPSTORM_META\type;
             if($data[0]->Role == "outletmanager"){
                 $outlets = new Outlet();
                 $outlet = $outlets->where("Manager", $data[0]->EmployeeNo);
-                $data[0]->Outlet = $outlet[0]->OutletCode;
-                if($data[0]->Outlet == null){
+                
+                if($outlet == null){
                     echo $this->view("admin/editsystemuser", ["data" => $data]);
                 }
                 else{
-                   $this->redirect(BASE_URL."AdminControls/loadUsersView");
+                    $hasOutlet = "yes";
+                    echo $this->view("admin/editsystemuser", ["data" => $data, "hasOutlet" => $hasOutlet]);
                 }
             }else{
                 echo $this->view("admin/editsystemuser", ["data" => $data]);
             }
-            
         }
